% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svytldr.R
\name{svytldr}
\alias{svytldr}
\title{svytldr}
\usage{
svytldr(
  df,
  ids,
  strata,
  weights,
  svyitem,
  svygrp,
  fltr_refuse = T,
  fltr_nas = T,
  flg_low_n = F,
  wide = F
)
}
\arguments{
\item{df}{A survey dataframe consisting of at minimum a survey item formatted as a factor variable.}

\item{ids}{Survey case ids (optional)}

\item{strata}{Survey strata (optional)}

\item{weights}{Survey weights (optional)}

\item{svyitem}{A survey item with factor (or ordered factor) format}

\item{svygrp}{A survey grouping variable, can be binary or multiple group, in factor format (optional)}

\item{fltr_refuse}{Filter refusals formatted 'refused' (Default = TRUE)}

\item{fltr_nas}{Filter NAs across dataframe (Default = TRUE)}

\item{flg_low_n}{Flag estimates with less than n = 100 in either svyitem response option or svygroup (or the combination thereof)}

\item{wide}{Produces a formatted table with columns for each group and statistic (Default = F; statistics nested w/in group)}
}
\value{
A tibble with M, SE, and unweighted Ns for each response for svyitem (or each response for svyitem within svygroup)
}
\description{
svytldr
}
\examples{
svytldr(df = df, ids = id, strata = strata, weights = wt, svyitem = "svyitem", svygrp = "group")
}
